refactor(panel-header): 패널 페이지에서 헤더 컴포넌트로 분리한다 (#353)
