+++ src/pages/Panel.tsx
@@ -14,11 +14,8 @@ import {
   useLoaderData,
 } from 'react-router-dom';
 
-import { AccountActionMenu } from '@/components/home/AccountActionMenu';
-import { OpenActionMenuArea } from '@/components/system/OpenActionMenuArea';
-import { SheetController } from '@/components/system/SheetController';
-import { Logo, Account, Menu } from '@/components/vectors';
-import { useOverlay } from '@/hooks/useOverlay';
+import { PanelHeader } from '@/components/panel/PanelHeader';
+import { Logo } from '@/components/vectors';
 import { getPanel } from '@/lib/api/panel';
 import { ApiError } from '@/lib/apiClient';
 // [NOTE] 로더 성공 반환값은 any 혹은 null로 고정한다
@@ -46,45 +43,10 @@ export const panelLoader: LoaderFunction = async ({ params }) => {
 
 export function Panel(): JSX.Element {
   const panel = useLoaderData() as PanelData;
-  const overlay = useOverlay();
-
-  function handleMenuButtonClick(): void {
-    overlay.open(({ close }) => (
-      <SheetController aria-label="메뉴" close={close}>
-        메뉴
-      </SheetController>
-    ));
-  }
 
   return (
     <>
-      <header>
-        <button
-          type="button"
-          className="hover:bg-grey-light rounded-full p-2"
-          aria-label="메뉴 열기"
-          onClick={handleMenuButtonClick}
-        >
-          <div role="img" aria-label="메뉴 아이콘">
-            <Menu className="fill-grey-darkest" />
-          </div>
-        </button>
-        <h1>{panel.title}</h1>
-        <OpenActionMenuArea
-          aria-label="내 계정 액션 메뉴"
-          open={({ close }) => <AccountActionMenu close={close} />}
-        >
-          <button
-            type="button"
-            className="hover:bg-grey-light rounded-full p-2"
-            aria-label="내 계정 아이콘"
-          >
-            <div role="img" aria-label="내 계정 아이콘">
-              <Account className="fill-grey-darkest" />
-            </div>
-          </button>
-        </OpenActionMenuArea>
-      </header>
+      <PanelHeader panel={panel} />
     </>
   );
 }
